requirements.txt
Pillow==12.1.1
pyexiv2==2.15.5
